requirements.txt
pymodbus>=3.8.3
pyserial
pyyaml
typing
logging
